+++ app/Http/Controllers/Dashboard/CarController.php
@@ -219,16 +219,10 @@ class CarController extends Controller
      */
     public function destroy(Car $car)
     {
+
         $images=car_img::where("car_id",$car->id)->get();
         Car::unlink_img($images);
-        $CarBadges=car_badge::where('car_id', '=', $car->id)->get();
-        foreach($CarBadges as $key=>$badge){
-            $badge->delete();
-        }
-        $CarFeatures=car_feature::where('car_id', '=', $car->id)->get();
-        foreach($CarFeatures as $key=>$feature){
-            $feature->delete();
-        }
+
         $car->delete();
         session()->flash('deleted',__("Changes has been Deleted Successfully"));
         return redirect()->route("dashboard.car.index");
@@ -237,31 +231,20 @@ class CarController extends Controller
         if (is_array(request('item'))) {
 			foreach (request('item') as $id) {
                 $car = Car::find($id);
-
+                if($car){
                 $images=car_img::where("car_id",$car->id)->get();
                 Car::unlink_img($images);
-                $CarBadges=car_badge::where('car_id', '=', $car->id)->get();
-                foreach($CarBadges as $key=>$badge){
-                    $badge->delete();
-                }
-                $CarFeatures=car_feature::where('car_id', '=', $car->id)->get();
-                foreach($CarFeatures as $key=>$feature){
-                    $feature->delete();
-                }
                 $car->delete();
+                }
+
 			}
 		} else {
             $car = Car::find(request('item'));
-            $images=car_img::where("car_id",$car->id)->get();
-            Car::unlink_img($images);
-            $CarBadges=car_badge::where('car_id', '=', $car->id)->get();
-            foreach($CarBadges as $key=>$badge){
-                $badge->delete();
+            if($car){
+                $images=car_img::where("car_id",$car->id)->get();
+                Car::unlink_img($images);
             }
-            $CarFeatures=car_feature::where('car_id', '=', $car->id)->get();
-            foreach($CarFeatures as $key=>$feature){
-                $feature->delete();
-            }
+
             $car->delete();
 		}
         session()->flash('deleted',__("Changes has been Deleted Successfully"));
